+++ app.js
@@ -18,6 +18,63 @@ const connector = new builder.ChatConnector({
     appPassword: process.env.MICROSOFT_APP_PASSWORD
 });
 
-assessBot.create(connector);
-
+const bot = new builder.UniversalBot(connector)
 server.post('/api/messages', connector.listen());
+
+// // Luis and intent recognition magic
+// const model = `https://api.projectoxford.ai/luis/v1/application?id=${process.env.LUIS_ID}&subscription-key=${process.env.LUIS_KEY}&verbose=true`
+//
+// bot.recognizer(new builder.LuisRecognizer(model))
+
+const scriptChoices = [
+  "demo script 1",
+  "demo script 2"
+];
+
+  // Send proactive messages on connection
+bot.on('conversationUpdate', (message) => {
+  if (message.membersAdded) {
+    if (message.membersAdded[0].id == 'default-bot') {
+        const initialProactiveMessage = new builder.Message()
+            .address(message.address)
+            .text("Welcome to the Kids Help Phone Assessment Bot!");
+        bot.send(initialProactiveMessage);
+
+        const secondProactiveMessage = new builder.Message()
+              .address(message.address)
+              .text("Type 'scripts' for a list of scripts to run");
+          bot.send(secondProactiveMessage);
+    }
+  }
+});
+
+// Send script list
+bot.dialog('/', ((session) => {
+  var msg = new builder.Message(session)
+  	.text("Please select a script to run:")
+  	.suggestedActions(
+  		builder.SuggestedActions.create(
+  				session, [
+  					builder.CardAction.imBack(session, "demo script 1", "demo script 1"),
+  					builder.CardAction.imBack(session, "demo script 2", "demo script 2")
+  				]
+  			));
+  session.send(msg);
+})).triggerAction({ matches: /^scripts/i });
+
+
+// Call createLibrary for all scripts
+bot.library(require('./scripts/demo1').createLibrary());
+bot.library(require('./scripts/demo2').createLibrary());
+const awardPoints = require('./scripts/scores');
+
+// Middleware for logging scores
+bot.use({
+    receive: function (event, next) {
+
+      // // conditionally award points for each answer
+      awardPoints(event.text);
+
+      next();
+    }
+});
